Removed XHR definition from module level global.
A var called 'XHR' was created at global scope within the module.  This means that using this in an environment where you want to replace XMLHttpRequest with a fake version at random intervals (i.e. when unit testing using Sinon.js or other fake server implementations).  You may not want to turn the fake server on too early (i.e. before this code loads) as this may prevent XHR requests for loading other modules from working.
So, this change simply re defines 'XHR' in the 2 functions where it is used, therefore it can be stubbed.